requirements.txt
# Core dependencies
streamlit>=1.25.0
pandas>=1.5.3
matplotlib>=3.7.2
networkx>=3.1
python-dotenv>=1.0.0
requests>=2.28.2
pytz>=2023.3

# Vector database and embeddings
faiss-cpu>=1.7.4
scipy>=1.10.1
numpy>=1.24.3
scikit-learn>=1.2.2

# Data processing and visualization
plotly>=5.14.0
pydeck>=0.8.0

# For local development
ipykernel>=6.22.0
